Refactor and split in files
